Parsing function calls with dot expressions as identifiers correctly.
They where previously parsed like this: a.(b()), where the root is a dot expression with a call expression as child.
They are now pared like this: (a.b)(), where the root is a call expression, and the identifier of the call expression is a dot expression.
Children of dot expressions can only be identifiers now.