Debuggin a crash happening in release builds.
The release builds crash when executing bytecode. From what I've seen it
always seems to crash in print_num(). GDB shows we are trying to write
to an invalid address, and valgrind (on the gcc build) enforces this.
When running valgrind on the clang build the crash happens later, but
seems to caused by the same problem (uninitialized memory errors).

The next step is probably to guard all access to the interpreters stack,
like we did in previous versions of the bytecode. This wil hopefully
help located the instruction or other source of the problem (if that is
actuall the problem).